requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.38.0
google-auth>=2.23.0
pydantic>=2.5.0
httpx>=0.25.0
python-dotenv>=1.0.0
sqlmodel>=0.0.14
apscheduler>=3.10.0
Pillow>=10.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0
